+++ zc-web-spring-boot-starter/src/main/java/com/zhengcheng/web/util/IpAddressUtils.java
@@ -1,5 +1,6 @@
 package com.zhengcheng.web.util;
 
+import com.zhengcheng.common.constant.CommonConstants;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
@@ -11,18 +12,19 @@ import javax.servlet.http.HttpServletRequest;
  * @date :    2019/10/6 20:26
  */
 public class IpAddressUtils {
+
     /**
      * 获取Ip地址
      *
-     * @param request
-     * @return
+     * @param request 请求
+     * @return IP
      */
     public static String getIpAddress(HttpServletRequest request) {
         String Xip = request.getHeader("X-Real-IP");
         String XFor = request.getHeader("X-Forwarded-For");
-        if (StringUtils.isNotEmpty(XFor) && !"unKnown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isNotEmpty(XFor) && !CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             //多次反向代理后会有多个ip值，第一个ip才是真实ip
-            int index = XFor.indexOf(",");
+            int index = XFor.indexOf(CommonConstants.COMMA);
             if (index != -1) {
                 return XFor.substring(0, index);
             } else {
@@ -30,22 +32,22 @@ public class IpAddressUtils {
             }
         }
         XFor = Xip;
-        if (StringUtils.isNotEmpty(XFor) && !"unKnown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isNotEmpty(XFor) && !CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             return XFor;
         }
-        if (StringUtils.isBlank(XFor) || "unknown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isBlank(XFor) || CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             XFor = request.getHeader("Proxy-Client-IP");
         }
-        if (StringUtils.isBlank(XFor) || "unknown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isBlank(XFor) || CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             XFor = request.getHeader("WL-Proxy-Client-IP");
         }
-        if (StringUtils.isBlank(XFor) || "unknown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isBlank(XFor) || CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             XFor = request.getHeader("HTTP_CLIENT_IP");
         }
-        if (StringUtils.isBlank(XFor) || "unknown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isBlank(XFor) || CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             XFor = request.getHeader("HTTP_X_FORWARDED_FOR");
         }
-        if (StringUtils.isBlank(XFor) || "unknown".equalsIgnoreCase(XFor)) {
+        if (StringUtils.isBlank(XFor) || CommonConstants.UNKNOWN.equalsIgnoreCase(XFor)) {
             XFor = request.getRemoteAddr();
         }
         return XFor;