+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - 技术分享 [API接口防止参数篡改和重放攻击](https://note.youdao.com/ynoteshare1/index.html?id=ed15f29e7ad1ff2d15a1236231283bc7&type=note)
 - 删除 `FeignInterceptorConfig` , 默认配置中增加 Feign OAuth2 拦截器
 - `TraceIdInterceptor` 增加 `applicationName` 构造参数，修复无法打印项目名的问题
+- [Java 两级缓存框架](https://gitee.com/ld/J2Cache) 的深入研究
 
 ## [v4.3.0] 2020.4.18